# openapi/shorter_dom_gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen DO NOT EDIT.
package domaim

// Error defines model for Error.
type Error struct {

	// Error code
	Code int32 `json:"code"`

	// Error message
	Message string `json:"message"`
}

// MakeUrlHash defines model for MakeUrlHash.
type MakeUrlHash struct {
	Url string `json:"Url"`
}

// UrlHash defines model for UrlHash.
type UrlHash struct {
	Hash string `json:"Hash"`
	Url  string `json:"Url"`
}

// MakeUrlHashJSONBody defines parameters for MakeUrlHash.
type MakeUrlHashJSONBody MakeUrlHash

// MakeUrlHashJSONRequestBody defines body for MakeUrlHash for application/json ContentType.
type MakeUrlHashJSONRequestBody MakeUrlHashJSONBody